Creation of all entities

+++ back/ws/src/main/java/com/recipebook/ws/repository/RecipeRepository.java
@@ -3,7 +3,7 @@ package com.recipebook.ws.repository;
 import com.recipebook.ws.entity.Recipe;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecipeRepository extends JpaRepository<Recipe, Integer> {
+public interface RecipeRepository extends JpaRepository<Recipe, Long> {
 
     Recipe findById(int id);
 }
